removing token infra from infra deployment
