+++ deployments/local/usdc/deploy.ts
@@ -1,5 +1,5 @@
 import { Deployed, DeploymentManager } from '../../../plugins/deployment_manager';
-import { DeploySpec, deployComet } from '../../../src/deploy';
+import { DeploySpec, deployComet} from '../../../src/deploy';
 
 export default async function deploy(deploymentManager: DeploymentManager, deploySpec: DeploySpec): Promise<Deployed> {
   // Set verification strategy to none to skip contract verification
@@ -19,4 +19,99 @@ export default async function deploy(deploymentManager: DeploymentManager, deplo
 
 
   return { ...deployed, ...infrastructureContracts };
+}
+
+import { FaucetToken, SimplePriceFeed } from '../../../build/types';
+import { exp, wait } from '../../../src/deploy';
+
+// Helper function to create tokens (same pattern as other deployment scripts)
+async function makeToken(
+  deploymentManager: DeploymentManager,
+  symbol: string,
+  name: string,
+  decimals: number
+): Promise<FaucetToken> {
+  const mint = (BigInt(1000000) * 10n ** BigInt(decimals)).toString();
+  return deploymentManager.deploy(symbol, 'test/FaucetToken.sol', [mint, name, decimals, symbol]);
+}
+
+async function makePriceFeed(
+  deploymentManager: DeploymentManager,
+  alias: string,
+  initialPrice: number,
+  decimals: number
+): Promise<SimplePriceFeed> {
+  return deploymentManager.deploy(alias, 'test/SimplePriceFeed.sol', [initialPrice * decimals, decimals]);
+}
+
+async function deployTokenInfra(deploymentManager: DeploymentManager){
+
+  // Deploy shared admin and governance contracts
+  const trace = deploymentManager.tracer();
+  const admin = await deploymentManager.getSigner();
+
+  const fauceteer = await deploymentManager.deploy('fauceteer', 'test/Fauceteer.sol', []);
+
+  // Deploy test tokens
+  const DAI = await makeToken(deploymentManager, 'DAI', 'DAI', 18);
+  const USDC = await makeToken(deploymentManager, 'USDC', 'USDC', 6);
+  const WETH = await makeToken(deploymentManager, 'WETH', 'Wrapped Ether', 18);
+  const WBTC = await makeToken(deploymentManager, 'WBTC', 'Wrapped Bitcoin', 8);
+  const LINK = await makeToken(deploymentManager, 'LINK', 'Chainlink', 18);
+  const UNI = await makeToken(deploymentManager, 'UNI', 'Uniswap', 18);
+
+  // Deploy price feeds using makePriceFeed function
+  const daiPriceFeed = await makePriceFeed(deploymentManager, 'daiPriceFeed', 1, 8); // $1.00 price
+  const usdcPriceFeed = await makePriceFeed(deploymentManager, 'usdcPriceFeed', 1, 8); // $1.00 price
+  const wethPriceFeed = await makePriceFeed(deploymentManager, 'wethPriceFeed', 2000, 8); // $2000 price
+  const wbtcPriceFeed = await makePriceFeed(deploymentManager, 'wbtcPriceFeed', 40000, 8); // $40000 price
+  const compPriceFeed = await makePriceFeed(deploymentManager, 'compPriceFeed', 50, 8); // $50 price
+  const linkPriceFeed = await makePriceFeed(deploymentManager, 'linkPriceFeed', 15, 8); // $15 price
+  const uniPriceFeed = await makePriceFeed(deploymentManager, 'uniPriceFeed', 10, 8); // $10 price
+
+  trace(`Attempting to mint tokens to fauceteer as ${admin.address}...`);
+
+  // Mint tokens to fauceteer
+  const tokenConfigs = [
+    { token: DAI, units: 1e8, name: 'DAI' },
+    { token: USDC, units: 1e6, name: 'USDC' },
+    { token: WETH, units: 1e6, name: 'WETH' },
+    { token: WBTC, units: 1e4, name: 'WBTC' },
+    { token: LINK, units: 1e7, name: 'LINK' },
+    { token: UNI, units: 1e7, name: 'UNI' },
+  ];
+
+  await Promise.all(
+    tokenConfigs.map(({ token, units, name }) => {
+      return deploymentManager.idempotent(
+        async () => (await token.balanceOf(fauceteer.address)).eq(0),
+        async () => {
+          trace(`Minting ${units} ${name} to fauceteer`);
+          const amount = exp(units, await token.decimals());
+          trace(await wait(token.connect(admin).allocateTo(fauceteer.address, amount)));
+          trace(`token.balanceOf(${fauceteer.address}): ${await token.balanceOf(fauceteer.address)}`);
+        }
+      );
+    })
+  );
+
+  return {
+    fauceteer,
+    // Tokens
+    DAI,
+    USDC,
+    WETH,
+    WBTC,
+    LINK,
+    UNI,
+    
+    // Price Feeds
+    daiPriceFeed,
+    usdcPriceFeed,
+    wethPriceFeed,
+    wbtcPriceFeed,
+    compPriceFeed,
+    linkPriceFeed,
+    uniPriceFeed,
+  }
 }
